+++ pkg/boxesimpl/boxesimpl.go
@@ -46,6 +46,7 @@ func DrawBoxesFromFile(inputFile, outputFile string) error {
 	doc.ConnectBoxes()
 	doc.AdjustDocHeightTLegend(textDimensionCalulator)
 	doc.IncludeComments(textDimensionCalulator)
+	doc.IncludeOverlays(textDimensionCalulator)
 	output, err := os.Create(outputFile)
 	svgdrawing := svgdrawing.NewDrawing(output)
 	svgdrawing.Start(doc.Title, doc.Height, doc.Width)
@@ -53,6 +54,7 @@ func DrawBoxesFromFile(inputFile, outputFile string) error {
 	svgdrawing.DrawRaster(doc.Width, doc.Height, types.RasterSize)
 	doc.DrawBoxes(svgdrawing)
 	doc.DrawConnections(svgdrawing)
+	doc.DrawOverlays(svgdrawing, textDimensionCalulator)
 	doc.DrawTitle(svgdrawing, textDimensionCalulator)
 	doc.DrawLegend(svgdrawing, textDimensionCalulator)
 	doc.DrawComments(svgdrawing, textDimensionCalulator)
@@ -83,24 +85,21 @@ func isRelatedToId(b boxes.Layout, filter []string) bool {
 	return false
 }
 
-func collectTruncatedFromCont(b []boxes.Layout, newId string, truncated *map[string]TruncatedInfo, hideComments bool) {
+func collectTruncatedFromCont(b []boxes.Layout, newId string, truncated *map[string]TruncatedInfo) {
 	for _, l := range b {
-		collectTruncated(l, newId, truncated, hideComments)
+		collectTruncated(l, newId, truncated)
 	}
 }
 
-func collectTruncated(b boxes.Layout, newId string, truncated *map[string]TruncatedInfo, hideComments bool) {
+func collectTruncated(b boxes.Layout, newId string, truncated *map[string]TruncatedInfo) {
 	if b.Id != "" {
-		if hideComments && b.Comment != nil {
-			b.Comment = nil
-		}
 		(*truncated)[b.Id] = TruncatedInfo{
 			truncated: b,
 			newId:     newId,
 		}
 	}
-	collectTruncatedFromCont(b.Horizontal, newId, truncated, hideComments)
-	collectTruncatedFromCont(b.Vertical, newId, truncated, hideComments)
+	collectTruncatedFromCont(b.Horizontal, newId, truncated)
+	collectTruncatedFromCont(b.Vertical, newId, truncated)
 }
 
 type TruncatedInfo struct {
@@ -119,31 +118,15 @@ func getNewId() string {
 	return fmt.Sprintf("xxx_%d", globalId)
 }
 
-func removeCommentsInCase(b boxes.Layout, hideComments bool) boxes.Layout {
-	if hideComments {
-		if b.Comment != nil {
-			b.Comment = nil
-		}
-		for i := range b.Connections {
-			c := b.Connections[i]
-			if c.Comment != nil {
-				c.Comment = nil
-			}
-		}
-	}
-	return b
-}
-
-func truncBoxes(b boxes.Layout, currentDepth, maxDepth int, expanded, blacklisted []string, hideComments bool) (boxes.Layout, map[string]TruncatedInfo) {
+func truncBoxes(b boxes.Layout, currentDepth, maxDepth int, expanded, blacklisted []string) (boxes.Layout, map[string]TruncatedInfo) {
 	truncatedBoxes := make(map[string]TruncatedInfo, 0)
-	b = removeCommentsInCase(b, hideComments)
 	if (currentDepth >= maxDepth) && (!b.Expand) && (!isRelatedToId(b, expanded)) {
 		// possible removed connections to this object
 		if b.Id == "" {
 			b.Id = getNewId()
 		}
-		collectTruncatedFromCont(b.Horizontal, b.Id, &truncatedBoxes, hideComments)
-		collectTruncatedFromCont(b.Vertical, b.Id, &truncatedBoxes, hideComments)
+		collectTruncatedFromCont(b.Horizontal, b.Id, &truncatedBoxes)
+		collectTruncatedFromCont(b.Vertical, b.Id, &truncatedBoxes)
 		b.Horizontal = make([]boxes.Layout, 0)
 		b.Vertical = make([]boxes.Layout, 0)
 	} else {
@@ -156,10 +139,10 @@ func truncBoxes(b boxes.Layout, currentDepth, maxDepth int, expanded, blackliste
 					if b.Id == "" {
 						b.Id = getNewId()
 					}
-					collectTruncated(b.Horizontal[i], b.Id, &truncatedBoxes, hideComments)
+					collectTruncated(b.Horizontal[i], b.Id, &truncatedBoxes)
 					continue
 				}
-				l, trunc := truncBoxes(b.Horizontal[i], currentDepth+1, maxDepth, expanded, blacklisted, hideComments)
+				l, trunc := truncBoxes(b.Horizontal[i], currentDepth+1, maxDepth, expanded, blacklisted)
 				addTruncated(trunc, &truncatedBoxes)
 				cont = append(cont, l)
 			}
@@ -174,10 +157,10 @@ func truncBoxes(b boxes.Layout, currentDepth, maxDepth int, expanded, blackliste
 					if b.Id == "" {
 						b.Id = getNewId()
 					}
-					collectTruncated(b.Vertical[i], b.Id, &truncatedBoxes, hideComments)
+					collectTruncated(b.Vertical[i], b.Id, &truncatedBoxes)
 					continue
 				}
-				l, trunc := truncBoxes(b.Vertical[i], currentDepth+1, maxDepth, expanded, blacklisted, hideComments)
+				l, trunc := truncBoxes(b.Vertical[i], currentDepth+1, maxDepth, expanded, blacklisted)
 				addTruncated(trunc, &truncatedBoxes)
 				cont = append(cont, l)
 			}
@@ -209,7 +192,7 @@ func copyTruncatedComments(layout *boxes.Layout, truncatedObjects map[string]Tru
 	}
 }
 
-func copyTruncatedConnections(layout *boxes.Layout, truncatedObjects map[string]TruncatedInfo, hideComments bool) {
+func copyTruncatedConnections(layout *boxes.Layout, truncatedObjects map[string]TruncatedInfo) {
 	// copy all needed connections from truncated objects to the current object
 	for _, v := range truncatedObjects {
 		if v.newId == layout.Id {
@@ -226,10 +209,14 @@ func copyTruncatedConnections(layout *boxes.Layout, truncatedObjects map[string]
 				if !connectionExistsByDestId(layout.Connections, destIdToUse) {
 					if destIdToUse != layout.Id {
 						c.DestId = destIdToUse
-						if hideComments && c.Comment != nil {
-							c.Comment = nil
-						}
 						(*layout).Connections = append(layout.Connections, c)
+					}
+				} else if c.Comment != nil {
+					for i := range layout.Connections {
+						con := &layout.Connections[i]
+						if con.DestId == destIdToUse {
+							con.HiddenComments = true
+						}
 					}
 				}
 			}
@@ -250,34 +237,30 @@ func adjustDestIdInRespectOfTruncated(layout *boxes.Layout, truncatedObjects map
 	layout.Connections = normalizedConnections
 }
 
-func adjustTruncated(layout *boxes.Layout, truncatedObjects map[string]TruncatedInfo, hideComments bool) {
-	copyTruncatedConnections(layout, truncatedObjects, hideComments)
+func adjustTruncated(layout *boxes.Layout, truncatedObjects map[string]TruncatedInfo) {
+	copyTruncatedConnections(layout, truncatedObjects)
 	copyTruncatedComments(layout, truncatedObjects)
 	adjustDestIdInRespectOfTruncated(layout, truncatedObjects)
 	if len(layout.Horizontal) > 0 {
-		adjustTruncatedForCont(&layout.Horizontal, truncatedObjects, hideComments)
+		adjustTruncatedForCont(&layout.Horizontal, truncatedObjects)
 	}
 	if len(layout.Vertical) > 0 {
-		adjustTruncatedForCont(&layout.Vertical, truncatedObjects, hideComments)
+		adjustTruncatedForCont(&layout.Vertical, truncatedObjects)
 	}
 }
 
-func adjustTruncatedForCont(cont *[]boxes.Layout, truncatedObjects map[string]TruncatedInfo, hideComments bool) {
+func adjustTruncatedForCont(cont *[]boxes.Layout, truncatedObjects map[string]TruncatedInfo) {
 	if cont != nil {
 		for i := range len(*cont) {
-			adjustTruncated(&(*cont)[i], truncatedObjects, hideComments)
+			adjustTruncated(&(*cont)[i], truncatedObjects)
 		}
 	}
 }
 
 func FilterBoxes(layout boxes.Boxes, defaultDepth int, expanded, blacklisted []string) boxes.Boxes {
-	return FilterBoxesComments(layout, defaultDepth, expanded, blacklisted, false)
-}
-
-func FilterBoxesComments(layout boxes.Boxes, defaultDepth int, expanded, blacklisted []string, hideComments bool) boxes.Boxes {
 	filteredBoxes := boxes.CopyBoxes(&layout)
-	b, truncatedObjects := truncBoxes(layout.Boxes, 0, defaultDepth, expanded, blacklisted, hideComments)
-	adjustTruncated(&b, truncatedObjects, hideComments)
+	b, truncatedObjects := truncBoxes(layout.Boxes, 0, defaultDepth, expanded, blacklisted)
+	adjustTruncated(&b, truncatedObjects)
 	filteredBoxes.Boxes = b
 	return *filteredBoxes
 }
@@ -300,14 +283,17 @@ func DrawBoxesFiltered(layout boxes.Boxes, defaultDepth int, expanded, blacklist
 
 func DrawBoxesFilteredComments(layout boxes.Boxes, defaultDepth int, expanded, blacklisted []string, hideComments, debug bool) UIReturn {
 	textDimensionCalulator := svgdrawing.NewSvgTextDimensionCalculator()
-	filteredLayout := FilterBoxesComments(layout, defaultDepth, expanded, blacklisted, hideComments)
+	filteredLayout := FilterBoxes(layout, defaultDepth, expanded, blacklisted)
 	doc, err := InitialLayoutBoxes(&filteredLayout, textDimensionCalulator)
 	if err != nil {
 		return UIReturn{ErrorMsg: fmt.Sprintf("error while initialy layout: %v", err)}
 	}
 	doc.ConnectBoxes()
 	doc.AdjustDocHeightTLegend(textDimensionCalulator)
-	doc.IncludeComments(textDimensionCalulator)
+	if !hideComments {
+		doc.IncludeComments(textDimensionCalulator)
+	}
+	doc.IncludeOverlays(textDimensionCalulator)
 	var svgBuilder strings.Builder
 	svgdrawing := svgdrawing.NewDrawing(&svgBuilder)
 	svgdrawing.Start(doc.Title, doc.Height, doc.Width)
@@ -325,6 +311,7 @@ func DrawBoxesFilteredComments(layout boxes.Boxes, defaultDepth int, expanded, b
 		doc.Boxes.DrawTextBoxes(svgdrawing)
 	}
 	doc.DrawConnections(svgdrawing)
+	doc.DrawOverlays(svgdrawing, textDimensionCalulator)
 	doc.DrawTitle(svgdrawing, textDimensionCalulator)
 	doc.DrawLegend(svgdrawing, textDimensionCalulator)
 	doc.DrawComments(svgdrawing, textDimensionCalulator)