+++ pkg/types/boxes/externals.go
@@ -33,8 +33,13 @@ type BoxesFileMixings struct {
     // Set of formats that overwrites the style of boxes, if specific conditions are met
     FormatVariations *FormatVariations  `yaml:"formatVariations,omitempty"`
 
+    // dictionary of comment objects, this comment will applied on layout objects and replace existing comments there
+    Comments map[string]types.Comment  `yaml:"comments,omitempty"`
+
     // optional map of images used in the generated graphic
     Images map[string]types.ImageDef  `yaml:"images,omitempty"`
+
+    Overlays []Overlay  `yaml:"overlays,omitempty"`
 }
 
 func NewBoxesFileMixings() *BoxesFileMixings {
@@ -44,7 +49,9 @@ func NewBoxesFileMixings() *BoxesFileMixings {
         Connections: make(map[string]ConnectionCont, 0),
         Formats: make(map[string]Format, 0),
         FormatVariations: NewFormatVariations(),
+        Comments: make(map[string]types.Comment, 0),
         Images: make(map[string]types.ImageDef, 0),
+        Overlays: make([]Overlay, 0),
     }
 }
 
@@ -69,9 +76,17 @@ func CopyBoxesFileMixings(src *BoxesFileMixings) *BoxesFileMixings {
         ret.Formats[k] = v
     }
     ret.FormatVariations = CopyFormatVariations(src.FormatVariations)
+    ret.Comments = make(map[string]types.Comment, 0)
+    for k, v := range src.Comments {
+        ret.Comments[k] = v
+    }
     ret.Images = make(map[string]types.ImageDef, 0)
     for k, v := range src.Images {
         ret.Images[k] = v
+    }
+    ret.Overlays = make([]Overlay, 0)
+    for _, e := range src.Overlays {
+        ret.Overlays = append(ret.Overlays, e)
     }
 
     return &ret
@@ -110,6 +125,9 @@ func CopyConnectionCont(src *ConnectionCont) *ConnectionCont {
 
     return &ret
 }
+
+
+