+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# v0.8.0
+## Boxes
+* improve comments on connections
+* add overlays
+
+# v0.7.0
+## Boxes
+* Comment mixins
+
 # v0.6.0
 ## Boxes
 * Extend the UI and the rendering engine to hide comments. The comments are